final shuffling, renaming, clean up, etc
moved dev under test, and got rid of dev build in favor of using dist/cedar.js
updated npm scripts accordingly, main ones are: build, test, coverage, start, predeploy
moved karma configuration out of gruntfile and into karma configs
pointed travis back at `npm test`, which first runs a build now
removed unused .min.js files - finally
updated README w/ dev instructions
updated CHANGELOG
