+++ karma.sauce.conf.js
@@ -1,0 +1,42 @@
+// import base configuration
+var base = require('./karma.conf');
+
+module.exports = function(config) {
+  var customLaunchers = {
+    sl_chrome: {
+      base: 'SauceLabs',
+      browserName: 'chrome',
+      platform: 'Windows 7',
+      version: '35'
+    },
+    sl_firefox: {
+      base: 'SauceLabs',
+      browserName: 'firefox',
+      version: '30'
+    },
+    sl_ios_safari: {
+      base: 'SauceLabs',
+      browserName: 'iphone',
+      platform: 'OS X 10.9',
+      version: '7.1'
+    },
+    sl_ie_11: {
+      base: 'SauceLabs',
+      browserName: 'internet explorer',
+      platform: 'Windows 8.1',
+      version: '11'
+    }
+  };
+  // apply base settings
+  base(config);
+  // then override
+  config.set({
+    sauceLabs: {
+      testName: 'Cedar Unit Tests'
+    },
+    customLaunchers: customLaunchers,
+    browsers: Object.keys(customLaunchers),
+    reporters: ['progress', 'saucelabs'],
+    singleRun: true
+  });
+};
