+++ package.json
@@ -4,21 +4,22 @@
   "description": "Visualization framework for the ArcGIS Platform",
   "main": "./dist/cedar.js",
   "scripts": {
-    "clean:dev": "rimraf ./dev/js",
+    "clean:coverage": "rimraf ./coverage",
     "clean:dist": "rimraf ./dist && mkdir dist",
     "clean:docs": "rimraf ./site/build",
-    "copy:dev": "copyfiles -f src/charts/*.json dev/js/charts",
     "copy:deps": "copyfiles -f node_modules/d3/*.js* site/build/js & copyfiles -f node_modules/vega/*.js* site/build/js",
     "copy:dist": "copyfiles -f src/charts/*.json dist/charts",
     "copy:docs": "copyfiles -f src/charts/*.json site/build/js/charts & copyfiles -f dist/*.js* site/build/js",
     "lint": "jshint src || true",
     "prebuild": "npm run lint && npm run clean:dist",
-    "build": "rollup -c profiles/dev.js & rollup -c profiles/dist.js & rollup -c profiles/prod.js & npm run copy:dist & npm run copy:dev",
-    "postbuild": "npm run test",
+    "build": "rollup -c profiles/dev.js & rollup -c profiles/prod.js & npm run copy:dist",
+    "pretest": "npm run build",
     "test": "karma start",
+    "precoverage": "npm run clean:coverage & npm run build",
+    "coverage": "karma start karma.coverage.conf.js",
     "prestart": "npm run clean:docs && npm run copy:deps",
-    "start": "watch 'npm run build && npm run copy:docs' src & grunt",
-    "predeploy": "npm run prestart && npm run build && npm run copy:docs"
+    "start": "watch 'npm run build && karma start && npm run copy:docs' src & grunt",
+    "predeploy": "npm run prestart && npm run test && npm run copy:docs"
   },
   "repository": {
     "type": "git",
@@ -68,7 +69,6 @@
     "mocha": "^3.0.1",
     "rimraf": "^2.5.4",
     "rollup": "^0.36.0",
-    "rollup-plugin-babel": "^2.6.1",
     "rollup-plugin-buble": "^0.14.0",
     "rollup-plugin-json": "^2.0.2",
     "rollup-plugin-uglify": "^1.0.1",
