--- profiles/dist.js
@@ -1,6 +1,0 @@
-import config from './base';
-
-config.dest = 'dist/cedar.js';
-config.sourceMap = 'dist/cedar.js.map';
-
-export default config;
