+++ test/pages/bar.html
@@ -1,6 +1,6 @@
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/d3/3.5.6/d3.js"></script>
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/vega/2.6.1/vega.js"></script>
-<script type="text/javascript" src="../js/cedar.js"></script>
+<script type="text/javascript" src="../../dist/cedar.js"></script>
 
 <div id="chart"></div>
